+++ src/components/sidebar/_drought.tsx
@@ -1,0 +1,5 @@
+import styles from "@/styles/SubBlocks.module.css";
+
+export default function Drought() {
+  return <div>Drought</div>
+}
