--- src/styles/FlowInfo.module.css
@@ -1,218 +1,0 @@
-.flowInfo {
-  position: absolute;
-  z-index: var(--z-flow-info);
-  top: 10px;
-  left: 40px;
-  filter: drop-shadow(0px 0px 25px rgba(41, 41, 41, 0.1));
-}
-
-.logoWrapper {
-  transform: scale(0.5) translateX(-50%);
-}
-
-.flowInfo > * {
-  padding: var(--padding);
-}
-
-.flowInfo > nav {
-  border-radius: var(--border-radius) var(--border-radius) 0 0;
-  background: var(--foreground);
-  color: var(--background);
-  padding-bottom: 0;
-}
-
-.content {
-  background: var(--background-80);
-  border-radius: 0 0 var(--border-radius) var(--border-radius);
-  padding-top: 30px;
-}
-
-.flowInfo > nav > button {
-  width: 100%;
-  text-align: left;
-  margin-bottom: 20px;
-}
-
-.flowInfo > nav > button > span {
-  font-size: var(--font-size-caption);
-  opacity: 0.5;
-  background: url("/icons/search.svg");
-  background-repeat: no-repeat;
-  padding: 2px 0 5px 20px;
-  width: 100%;
-  text-align: left;
-}
-
-.flowInfo > nav > button > h2 {
-  font-size: var(--font-size-title);
-  font-weight: var(--font-weight-bold);
-  margin: 8px 0 16px 0;
-}
-
-.flowInfo > nav > button:hover > span {
-  opacity: 1;
-}
-
-.flowInfo > nav > button:hover > h2 {
-  text-decoration: underline;
-}
-
-.tabBar > button {
-  margin-right: 20px;
-  color: var(--background);
-  padding: 6px 0 6px 32px;
-  font-weight: var(--font-weight-bold);
-  background-repeat: no-repeat;
-  background-size: 26px;
-  opacity: 0.5;
-}
-
-.tabBar > button:hover,
-.tabBar > button.selected {
-  opacity: 1;
-}
-
-.tabBar > button.selected {
-  border-bottom: 2px solid var(--background);
-}
-
-.consumer {
-  background: url("/icons/consumer.svg");
-}
-.producer {
-  background: url("/icons/producer.svg");
-}
-
-.summary {
-  display: flex;
-  border-bottom: 1px solid var(--foreground-10);
-  padding-bottom: 15px;
-}
-
-.stats {
-  padding-top: 15px;
-  border-bottom: 1px solid var(--foreground-10);
-}
-
-.stats > h3, .top > h3 {
-  font-size: var(--font-size-caption);
-  font-weight: var(--font-weight-normal);
-  margin-bottom: 10px;
-  color: var(--foreground-50);
-}
-
-.top {
-  padding-top: 15px;
-}
-
-
-.summary dt {
-  font-size: var(--font-size-caption);
-  color: var(--foreground-50);
-  margin-bottom: 5px;
-}
-
-.summary b {
-  font-size: var(--font-size-subtitle);
-  font-weight: var(--font-weight-bold);
-}
-
-.summary > dl:first-child {
-  margin-right: 15px;
-  width: 40%;
-  border-right: 1px solid var(--foreground-10);
-}
-
-.content {
-  max-height: 55vh;
-  overflow-x: hidden;
-  overflow-y: auto;
-}
-
-
-
-.crops > li {
-  margin-bottom: 30px;
-  cursor: pointer;
-}
-
-.crops dl {
-  display: flex;
-  justify-content: space-between;
-  position: relative;
-}
-
-.crops > li > dl > dt,
-.crops > li > dl > dd {
-  font-weight: var(--font-weight-bold);
-}
-
-.crops > li > dl::before,
-.crops > li > dl > dt::before,
-.crops > li > dl > dd::before {
-  height: 5px;
-  content: " ";
-  border-radius: 99px;
-}
-
-.crops > li > dl::before,
-.crops > li > dl > dd::before {
-  position: absolute;
-  display: block;
-  left: 0;
-  top: 20px;
-}
-
-.crops > li > dl::before {
-  width: 100%;
-  background: var(--background-2);
-}
-
-.crops > li > dl > dt::before {
-  display: inline-block;
-  width: 5px;
-  background: var(--color);
-  margin-right: 5px;
-  margin-bottom: 2px;
-}
-
-.crops > li > dl > dd::before {
-  width: var(--width);
-  background: var(--color);
-  transition: width 0.5s ease;
-}
-
-.loading .crops > li > dl > dd::before {
-  animation: fade 1s linear infinite;
-  width: 100%;
-}
-
-.detail {
-  margin-left: 30px;
-  margin-top: 25px;
-}
-
-.detail > li {
-  margin-bottom: 10px;
-}
-
-.detail dt::before {
-  display: inline-block;
-  width: 4px;
-  height: 4px;
-  content: " ";
-  border-radius: 99px;
-  background: var(--color);
-  margin-right: 5px;
-  margin-bottom: 2px;
-}
-
-@keyframes fade {
-  0%,
-  100% {
-    opacity: 0;
-  }
-  50% {
-    opacity: 1;
-  }
-}
