+++ src/styles/Summary.module.css
@@ -1,0 +1,20 @@
+.summary {
+  display: flex;
+}
+
+.summary dt {
+  font-size: var(--font-size-caption);
+  color: var(--foreground-50);
+  margin-bottom: 5px;
+}
+
+.summary b {
+  font-size: var(--font-size-subtitle);
+  font-weight: var(--font-weight-bold);
+}
+
+.summary > dl:first-child {
+  margin-right: 15px;
+  width: 40%;
+  border-right: 1px solid var(--foreground-10);
+}
